Add a new script that just lists files available for download with minimal information about them. Download ALL sets of files, not just the first set of files.  Any purchase that includes both video and audio files will have an array of arrays of files to download. Set STDOUT to UTF8 to avoid "wide character" complaints. Keep some statistics and report them, a prelude to not die-ing on failed downloads but instead continuing and reporting failures at the end. If a huge file already exists, check it.  Only skip it if it needs to be downloaded.  That allows one to manually download the huge file and then use the script to check its MD5.  Also actually report skipped files at the end. Append "mp4" to video downloads that end with 480p, 720p, or 1080p. In the file name, convert all UTF double-quotes to " and convert all UTF single-quotes to '
Add a new script that just lists files available for download with minimal information about them.
Download ALL sets of files, not just the first set of files.  Any purchase that includes both video and audio files will have an array of arrays of files to download.
Set STDOUT to UTF8 to avoid "wide character" complaints.
Keep some statistics and report them, a prelude to not die-ing on failed downloads but instead continuing and reporting failures at the end.
If a huge file already exists, check it.  Only skip it if it needs to be downloaded.  That allows one to manually download the huge file and then use the script to check its MD5.  Also actually report skipped files at the end.
Append "mp4" to video downloads that end with 480p, 720p, or 1080p.
In the file name, convert all UTF double-quotes to " and convert all UTF single-quotes to '